Add files via upload
AI MUSIC WEB APP - This Will be a modular project which will have 5 parts. You have to make an AI Music Web Application. This web application can change the song as per the wrist you show in front of the webcam. If you show your left wrist it will be playing a song, and if you show your right wrist it will play a different song.